refactor(core): split model extraction by domain
Why:
- The central extractor coupled document triage, paper facts, and objectives through response-model branching and obscured prompt and schema ownership.

What:
- Move provider calls, prompts, schemas, completion limits, retries, repairs, and traces into the document profile, paper fact, and objective domains.
- Keep only domain-neutral content and JSON normalization shared, and update direct callers, tests, benchmarks, and ownership docs.
- Remove the central extractor contract and its stale test naming without retaining compatibility wrappers.

Impact:
- Domain model contracts can evolve independently while HTTP, persistence, and pipeline behavior remain unchanged.
- The shared extraction package no longer owns provider or domain policy.

Tests:
- `./.venv/bin/python -m pytest tests/unit/services/test_domain_model_extractors.py tests/unit/services/test_domain_model_extraction_ownership.py -q` (109 passed)
- Full backend pytest coverage (effective 846 passed, 9 skipped after supplying the required settings-only `LENS_DATABASE_URL`)
- `python -m compileall -q application tests scripts/benchmarks`
- `python -c 'import main'`
- `python3 scripts/check_docs_governance.py`
- `pyflakes` on the new domain extraction production modules
- `git diff --cached --check`

Refs:
- n/a
