+++ backend/application/core/document_profiles/prompts.py
@@ -1,0 +1,25 @@
+from __future__ import annotations
+
+import json
+from typing import Any
+
+_DOCUMENT_PROFILE_SYSTEM_PROMPT = """
+You are doing document triage for a materials-literature backend.
+
+Non-negotiable rules:
+- This is coarse document classification, not knowledge extraction.
+- Return exactly one JSON object and nothing else.
+- Do not write natural-language summaries or explanations.
+- `doc_type` must be one of: experimental, review, mixed, uncertain.
+- `parsing_warnings` may only use: insufficient_content, classification_uncertain.
+- If the input is weak or ambiguous, return `uncertain`.
+""".strip()
+
+
+def build_document_profile_prompt(payload: dict[str, Any]) -> tuple[str, str]:
+    user_prompt = (
+        "Classify this document for lightweight Core document triage.\n\n"
+        f"Input JSON:\n{json.dumps(payload, ensure_ascii=False, indent=2)}\n\n"
+        "Return only schema-valid structured data. Do not add any explanation."
+    )
+    return _DOCUMENT_PROFILE_SYSTEM_PROMPT, user_prompt
