+++ backend/docs/plans/core/core-semantic-build-packaging-alignment-plan.md
@@ -152,16 +152,16 @@ collection-facing aggregators such as:
 - business-owned extraction contracts, prompt design, and response schema
 - infra-owned transport details
 
-The extraction contract belongs to Core semantic build, not to infra and not
-to a generic application client bag. The current transport implementation now
-lives directly in the owning Core extractor path:
+The extraction contracts belong to their Core domains, not to infra and not to
+a generic application client bag. The current transport implementations live
+directly in the owning domain paths:
 
-- [`../../../application/core/structured_extraction/extractor.py`](../../../application/core/structured_extraction/extractor.py)
+- [`../../../application/core/document_profiles/extraction.py`](../../../application/core/document_profiles/extraction.py)
+- [`../../../application/core/paper_facts/extraction.py`](../../../application/core/paper_facts/extraction.py)
+- [`../../../application/core/objectives/extraction.py`](../../../application/core/objectives/extraction.py)
 
-Therefore:
-
-- Core keeps `llm/prompts.py`, `llm/schemas.py`, and `llm/extractor.py`
-- Core currently also owns the concrete OpenAI-compatible transport call path
+Each domain keeps its own `prompts.py`, `schemas.py`, extraction policy, and
+concrete OpenAI-compatible transport call path.
 
 ## Ownership Boundary After Refactor
