+++ backend/scripts/benchmarks/text_window_probe.py
@@ -204,8 +204,8 @@ def load_text_window_components(
     backend_root: Path,
 ) -> tuple[Any, type[Any]]:
     ensure_backend_root_on_path(backend_root)
-    from application.core.structured_extraction.prompts import build_text_window_extraction_prompt
-    from application.core.structured_extraction.schemas import StructuredExtractionBundle
+    from application.core.paper_facts.prompts import build_text_window_extraction_prompt
+    from application.core.paper_facts.schemas import StructuredExtractionBundle
 
     return build_text_window_extraction_prompt, StructuredExtractionBundle
