+++ backend/tests/conftest.py
@@ -11,29 +11,29 @@ if str(ROOT) not in sys.path:
 
 
 @pytest.fixture(autouse=True)
-def _patch_core_llm_extractor(monkeypatch):
+def _patch_domain_model_extractors(monkeypatch):
     from application.core import comparison_service
     from application.core.document_profiles import (
         service as document_profile_service,
     )
     from application.core.objectives import research_objective_service
     from application.core.paper_facts import service as paper_facts_service
-    from tests.support.fake_core_llm_extractor import FakeCoreLLMStructuredExtractor
+    from tests.support.fake_domain_model_extractor import FakeDomainModelExtractor
 
-    fake = FakeCoreLLMStructuredExtractor()
+    fake = FakeDomainModelExtractor()
     monkeypatch.setattr(
         document_profile_service,
-        "build_default_core_llm_structured_extractor",
+        "build_default_document_profile_extractor",
         lambda: fake,
     )
     monkeypatch.setattr(
         paper_facts_service,
-        "build_default_core_llm_structured_extractor",
+        "build_default_paper_facts_extractor",
         lambda: fake,
     )
     monkeypatch.setattr(
         research_objective_service,
-        "build_default_core_llm_structured_extractor",
+        "build_default_objective_extractor",
         lambda: fake,
     )
